Both console and restapi updated

+++ empRestAPI/empRestAPI/Controllers/EmployeeController.cs
@@ -1,5 +1,6 @@
 ﻿using empRestAPI.Models;
 using empRestAPI.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -9,6 +10,7 @@ using System.Threading.Tasks;
 
 namespace empRestAPI.Controllers
 {
+    [Authorize]
     [Route("api/[controller]/[action]")]
     [ApiController]
     public class EmployeeController : ControllerBase
@@ -22,18 +24,25 @@ namespace empRestAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> AddEmployee(DataInput dataInput)
         {
+            if (string.IsNullOrWhiteSpace(dataInput.Department)) return BadRequest("Enter Department Name");
+            if (string.IsNullOrWhiteSpace(dataInput.EmployeeName)) return BadRequest("Enter Employee Name");
             var result = await employeeService.AddEmployee(dataInput);
             return Ok(result);
         }
         [HttpPut]
         public async Task<IActionResult> UpdateEmployee(DataInput dataInput)
         {
+            if (string.IsNullOrWhiteSpace(dataInput.Department)) return BadRequest("Enter Department Name");
+            if (string.IsNullOrWhiteSpace(dataInput.EmployeeName)) return BadRequest("Enter Employee Name");
+            if (dataInput.EmployeeID==0) return BadRequest("Enter Employee ID");
+
             var result = await employeeService.UpadteEmployee(dataInput);
             return Ok(result);
         }
         [HttpDelete("{employeeId}")]
         public async Task<IActionResult> DeleteEmployee(int employeeId)
         {
+            if (employeeId == 0) return BadRequest("Enter Employee ID");
             var result = await employeeService.DeleteEmployee(employeeId);
             return Ok(result);
         }
@@ -46,6 +55,7 @@ namespace empRestAPI.Controllers
         [HttpGet("{employeeId}")]
         public async Task<IActionResult> GetEmployee(int employeeId)
         {
+            if (employeeId == 0) return BadRequest("Enter Employee ID");
             var result = await employeeService.GetEmployee(employeeId);
             return Ok(result);
         }
